+++ packages/video_player/video_player/example/lib/main.dart
@@ -211,8 +211,8 @@ class _BumbleBeeRemoteVideoState extends State<_BumbleBeeRemoteVideo> {
   Future<ClosedCaptionFile> _loadCaptions() async {
     final String fileContents = await DefaultAssetBundle.of(context)
         .loadString('assets/bumble_bee_captions.vtt');
-    return WebVttCaptionFile(
-        fileContents); // For vtt files, use WebVttCaptionFile
+    return WebVTTCaptionFile(
+        fileContents); // For vtt files, use WebVTTCaptionFile
   }
 
   @override